+++ src/main.rs
@@ -1,6 +1,9 @@
 #![cfg_attr(all(windows, not(debug_assertions)), windows_subsystem = "windows")]
 
 #[cfg(windows)]
+mod updater;
+
+#[cfg(windows)]
 mod windows_app;
 
 #[cfg(windows)]